Regression test for getTimeZone.
The localtime_r call made from getTimeZone may or may not perform a tzset().
In particular, in glibc, a tzset() will only be performed the first time a process runs localtime_r.

This added regression test will fail on implementations like glibc that only
perform a tzset() on the first call to localtime_r.

A fix to make getTimeZone always call tzset() can be found in patch:

[Make getTimeZone cross-platform consistent by always considering the TZ environment variable.
oconnorr@google.com**20130610222254